backend/requirements.txt
Flask
Flask-CORS
python-dotenv
requests
beautifulsoup4
lxml
feedparser
